Add query base class and mutation base class
