+++ backend/src/PostService/PostService.Api/GraphQL/Mutation.cs
@@ -1,0 +1,126 @@
+using System.Security.Claims;
+using PostService.Application.Commands.AddComment;
+using PostService.Application.Commands.AddLike;
+using PostService.Application.Commands.AddPost;
+using PostService.Application.Commands.DeleteComment;
+using PostService.Application.Commands.DeleteLike;
+using PostService.Application.Commands.DeletePost;
+using PostService.Application.Commands.UpdatePost;
+using PostService.Application.DTOs;
+
+namespace PostService.Api.GraphQL;
+
+public class Mutation
+{
+    private readonly IHttpContextAccessor _httpContextAccessor;
+
+    public Mutation(IHttpContextAccessor httpContextAccessor)
+    {
+        _httpContextAccessor = httpContextAccessor;
+    }
+    
+    public async Task<PostDto> CreatePost(CreatePostDto input, [Service] AddPostCommandHandler addPostCommandHandler)
+    {
+        var userId = _httpContextAccessor.HttpContext?.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        var command = new AddPostCommand(input, userId);
+
+        var result = await addPostCommandHandler.HandleAsync(command);
+
+        if (!result.IsSuccess)
+        {
+            throw new GraphQLException(new Error(result.Error.Message));
+        }
+
+        return result.Response;
+    }
+
+    public async Task<PostDto> UpdatePost(UpdatePostDto input, [Service] UpdatePostCommandHandler updatePostCommandHandler)
+    {
+        var userId = _httpContextAccessor.HttpContext?.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        var command = new UpdatePostCommand(input, userId);
+
+        var result = await updatePostCommandHandler.HandleAsync(command);
+
+        if (!result.IsSuccess)
+        {
+            throw new GraphQLException(new Error(result.Error.Message));
+        }
+
+        return result.Response;
+    }
+
+    public async Task<string> DeletePost(Guid id, [Service] DeletePostCommandHandler deletePostCommandHandler)
+    {
+        var userId = _httpContextAccessor.HttpContext?.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        var command = new DeletePostCommand(id, userId);
+
+        var result = await deletePostCommandHandler.HandleAsync(command);
+
+        if (!result.IsSuccess)
+        {
+            throw new GraphQLException(new Error(result.Error.Message));
+        }
+
+        return result.Response;
+    }
+    
+    public async Task<CommentDto> CreateComment(CreateCommentDto input, [Service] AddCommentCommandHandler addCommentCommandHandler)
+    {
+        var userId = _httpContextAccessor.HttpContext?.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        var command = new AddCommentCommand(input, userId);
+
+        var result = await addCommentCommandHandler.HandleAsync(command);
+
+        if (!result.IsSuccess)
+        {
+            throw new GraphQLException(new Error(result.Error.Message));
+        }
+
+        return result.Response;
+    }
+    
+    public async Task<string> DeleteComment(Guid id, Guid postId, [Service] DeleteCommentCommandHandler deleteCommentCommandHandler)
+    {
+        var userId = _httpContextAccessor.HttpContext?.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        var command = new DeleteCommentCommand(id, postId, userId);
+
+        var result = await deleteCommentCommandHandler.HandleAsync(command);
+
+        if (!result.IsSuccess)
+        {
+            throw new GraphQLException(new Error(result.Error.Message));
+        }
+
+        return result.Response;
+    }
+    
+    public async Task<LikeDto> AddLike(Guid postId, [Service] AddLikeCommandHandler addLikeCommandHandler)
+    {
+        var userId = _httpContextAccessor.HttpContext?.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        var command = new AddLikeCommand(postId, userId);
+
+        var result = await addLikeCommandHandler.HandleAsync(command);
+
+        if (!result.IsSuccess)
+        {
+            throw new GraphQLException(new Error(result.Error.Message));
+        }
+
+        return result.Response;
+    }
+    
+    public async Task<string> DeleteLike(Guid id, Guid postId, [Service] DeleteLikeCommandHandler deleteLikeCommandHandler)
+    {
+        var userId = _httpContextAccessor.HttpContext?.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        var command = new DeleteLikeCommand(id, postId, userId);
+
+        var result = await deleteLikeCommandHandler.HandleAsync(command);
+
+        if (!result.IsSuccess)
+        {
+            throw new GraphQLException(new Error(result.Error.Message));
+        }
+
+        return result.Response;
+    }
+}
