+++ frontend/src/environments/environment.ts
@@ -1,5 +1,5 @@
 export const environment = {
   auth_service: "http://localhost:8000/auth/graphql",
-  post_service: "http://localhost:8000/post/",
+  post_service: "http://localhost:8000/post/graphql",
   production: false
 };
